# lib/api-zod/src/generated/types/checkInQrScanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInQrScanInput {
  /**
     * Opaque attendee QR payload; contains no attendee details or database identifiers.
     * @pattern ^shalom-checkin:v1:[A-Za-z0-9_-]{43}$
     */
  payload: string;
}